Yummers

+++ create-legacy-mod/src/main/java/com/melonstudios/createlegacy/util/INetworkLogger.java
@@ -1,6 +1,6 @@
 package com.melonstudios.createlegacy.util;
 
 public interface INetworkLogger {
-    void setSU(int su);
-    void setMaxSU(int su);
+    void setSU(float su);
+    void setMaxSU(float su);
 }
